refactor test_ask_search
